migrate watch,delete to new api
