+++ crates/fluvio-sc-schema/src/objects/delete.rs
@@ -5,32 +5,26 @@
 
 use std::fmt::Debug;
 
+use anyhow::Result;
+
 use fluvio_protocol::{Encoder, Decoder};
 use fluvio_protocol::api::Request;
 
 use crate::{DeletableAdminSpec};
 use crate::Status;
 use crate::AdminPublicApiKey;
-use super::{DeleteApiEnum, COMMON_VERSION};
-
-DeleteApiEnum!(DeleteRequest);
+use super::{COMMON_VERSION, TypeBuffer};
 
 // This can be auto generated by enum derive later
 #[derive(Debug, Default, Encoder, Decoder)]
-pub struct DeleteRequest<S: DeletableAdminSpec> {
-    key: S::DeleteKey,
-}
+pub struct ObjectApiDeleteRequest(TypeBuffer);
 
-impl<S> DeleteRequest<S>
-where
-    S: DeletableAdminSpec,
-{
-    pub fn new(key: S::DeleteKey) -> Self {
-        Self { key }
-    }
-
-    pub fn key(self) -> S::DeleteKey {
-        self.key
+impl ObjectApiDeleteRequest {
+    pub fn downcast<S>(&self) -> Result<Option<S::DeleteKey>>
+    where
+        S: DeletableAdminSpec,
+    {
+        self.0.downcast::<S, S::DeleteKey>()
     }
 }
